Fixed classloader errors and warnings (#82)
* Added jar-file-not-found message

* Removed set of scl field

* Removed useless RuntimeApplicationClassLoader

* Removed set of srl field and declaration method of environment file

* Added jar file of environment to classpath
Added try of finding environment file and using error-message

* Changed replaces of error message

* Marked ApplicationEnvironmentEvent as deprecated

* removed Locale.ENGLISH from toLoverCase() call

* Made ApplicationEnvironmentEvent unused because deprecation

* Loading all available jar files into the classpath

* Deprecate ApplicationEnvironmentEvent

* Make restart delay to a constant

* Remove unecessary import

+++ cloudnet/src/main/java/de/dytanic/cloudnet/service/JVMCloudService.java
@@ -24,6 +24,7 @@ import java.net.URL;
 import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
@@ -31,12 +32,15 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.UnaryOperator;
 import java.util.jar.JarFile;
+import java.util.stream.Collectors;
 
 final class JVMCloudService implements ICloudService {
 
     private static final String RUNTIME = "jvm";
 
     private static final String TEMP_NAME_SPLITTER = "_";
+    
+    private static final long SERVICE_ERROR_RESTART_DELAY = 10;
 
     private static final Lock START_SEQUENCE_LOCK = new ReentrantLock();
 
@@ -601,18 +605,45 @@ final class JVMCloudService implements ICloudService {
         ));
 
         File wrapperFile = new File(System.getProperty("cloudnet.tempDir", "temp"), "caches/wrapper.jar");
+        List<String> availableJarFiles = Files.list(this.directory.toPath())
+                .map(Path::toFile)
+                .filter(file -> file.getName().endsWith(".jar"))
+                .filter(file -> {
+                    for (ServiceEnvironment environment : this.serviceConfiguration.getProcessConfig().getEnvironment().getEnvironments()) {
+                        if (file.getName().toLowerCase()
+                                .contains(environment.getName().toLowerCase())) {
+                            return true;
+                        }
+                    }
+                    return false;
+                })
+                .map(File::getAbsolutePath)
+                .collect(Collectors.toList());
+
+        if (availableJarFiles.isEmpty()) {
+            System.out.println(LanguageManager.getMessage("cloud-service-jar-file-not-found-error")
+                    .replace("%time%", String.valueOf(SERVICE_ERROR_RESTART_DELAY))
+                    .replace("%environment%", this.serviceConfiguration.getProcessConfig().getEnvironment().name()
+                    ));
+            Thread.sleep(SERVICE_ERROR_RESTART_DELAY * 1000);
+            stop();
+            return;
+        }
 
         commandArguments.addAll(this.serviceConfiguration.getProcessConfig().getJvmOptions());
         commandArguments.addAll(Arrays.asList(
                 "-Xmx" + this.serviceConfiguration.getProcessConfig().getMaxHeapMemorySize() + "M",
                 "-javaagent:" + wrapperFile.getAbsolutePath(),
-                "-cp",
-                "\"" + System.getProperty("cloudnet.launcher.driver.dependencies") + File.pathSeparator + wrapperFile.getAbsolutePath() + "\""
+                "-cp", System.getProperty("cloudnet.launcher.driver.dependencies") +
+                        File.pathSeparator + wrapperFile.getAbsolutePath() +
+                        File.pathSeparator + String.join(File.pathSeparator, availableJarFiles)
         ));
 
         try (JarFile jarFile = new JarFile(wrapperFile)) {
             commandArguments.add(jarFile.getManifest().getMainAttributes().getValue("Main-Class"));
         }
+
+        commandArguments.add(availableJarFiles.get(0));
 
         this.postConfigureServiceEnvironmentStartParameters(commandArguments);
 
@@ -622,7 +653,7 @@ final class JVMCloudService implements ICloudService {
                 .start();
     }
 
-    private void postConfigureServiceEnvironmentStartParameters(List<String> commandArguments) throws Exception {
+    private void postConfigureServiceEnvironmentStartParameters(List<String> commandArguments) {
         switch (this.serviceConfiguration.getProcessConfig().getEnvironment()) {
             case MINECRAFT_SERVER:
                 commandArguments.add("nogui");
@@ -705,12 +736,10 @@ final class JVMCloudService implements ICloudService {
                 properties = new Properties();
 
                 file = new File(this.directory, "eula.txt");
-                if (file.exists()) {
+                if (file.exists() || file.createNewFile()) {
                     try (InputStream inputStream = new FileInputStream(file)) {
                         properties.load(inputStream);
                     }
-                } else {
-                    file.createNewFile();
                 }
 
                 properties.setProperty("eula", "true");
@@ -782,9 +811,7 @@ final class JVMCloudService implements ICloudService {
     }
 
     private void copyDefaultFile(String from, File target) throws Exception {
-        if (!target.exists()) {
-            target.createNewFile();
-
+        if (!target.exists() && target.createNewFile()) {
             try (InputStream inputStream = JVMCloudService.class.getClassLoader().getResourceAsStream(from);
                  OutputStream outputStream = new FileOutputStream(target)) {
                 if (inputStream != null) {
